Remove universal_config support for open_cache
