+++ cumulusci/core/config/project_config.py
@@ -5,6 +5,7 @@ import re
 from pathlib import Path
 from configparser import ConfigParser
 from itertools import chain
+from contextlib import contextmanager
 
 API_VERSION_RE = re.compile(r"^\d\d+\.0$")
 
@@ -30,6 +31,7 @@ from cumulusci.core.source import LocalFolderSource
 from cumulusci.core.source import NullSource
 from cumulusci.utils.git import current_branch, git_path
 from cumulusci.utils.yaml.cumulusci_yml import cci_safe_load
+from cumulusci.utils.fileutils import open_fs_resource
 
 from github3.exceptions import NotFoundError
 
@@ -827,3 +829,11 @@ class BaseProjectConfig(BaseTaskFlowConfig):
         cache_dir = Path(self.repo_root, ".cci")
         cache_dir.mkdir(exist_ok=True)
         return cache_dir
+
+    @contextmanager
+    def open_cache(self, cache_name):
+        "A context managed PyFilesystem-based cache which could theoretically be on any filesystem."
+        with open_fs_resource(self.cache_dir) as cache_dir:
+            cache = cache_dir / cache_name
+            cache.mkdir(exist_ok=True)
+            yield cache
