+++ cumulusci/core/config/universal_config.py
@@ -82,7 +82,3 @@ class UniversalConfig(BaseTaskFlowConfig):
                 "global_config": UniversalConfig.config_global,
             }
         )
-
-    @property
-    def cache_dir(self):
-        return self.config_global_local_path
